requirements.txt
Django==1.11.16
djangorestframework==3.9.0
mysqlclient==1.3.4
Pygments==2.2.0
pytz==2018.7
